+++ NEWS.md
@@ -1,3 +1,7 @@
-# semdplot 0.0.0.9001
+# semdplot 0.0.0.9002
 
-* WIP. Updates to be added.
+- **Work-In-Progress!**
+
+- The first crude draft, as a proof-of-concept. (0.0.0.9001)
+- Remove `GGally` from dependence and `plot_resid()`. Plots of
+  residuals are done internally using `ggplot2`. (0.0.0.9002)